split library and add doctests
